+++ .scalafix.conf
@@ -10,4 +10,3 @@ rules = [
 OrganizeImports {
   groupedImports = Merge
 }
-OrganizeImports.targetDialect = Scala3
